fix component structure
Previously made components were accommodated in the components folder
