+++ ReservacionDeVuelos/src/components/atoms/AddButton/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import {Text, TouchableOpacity} from 'react-native';
+import styles from './styles';
+
+export default function AddButton({navigation}) {
+  return (
+    <TouchableOpacity
+      style={styles.containerButton}
+      onPress={() => navigation.navigate('Passenger')}>
+      <Text style={styles.textButton}>+</Text>
+    </TouchableOpacity>
+  );
+}
