+++ ReservacionDeVuelos/src/components/atoms/BackButton/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {TouchableOpacity} from 'react-native';
+import BackSVG from '../../assets/icons/back.svg';
+import styles from './styles';
+
+export default function BackButton({navigation}) {
+  return (
+    <TouchableOpacity
+      style={styles.containerButton}
+      onPress={() => navigation.goBack()}>
+      <BackSVG />
+    </TouchableOpacity>
+  );
+}
