+++ ReservacionDeVuelos/src/components/atoms/BackButton/styles.js
@@ -1,0 +1,8 @@
+import {StyleSheet, Dimensions} from 'react-native';
+import colors from '../../consts/colors';
+
+const styles = StyleSheet.create({
+  containerButton: {marginTop: 9, marginLeft: 30},
+});
+
+export default styles;
