+++ ReservacionDeVuelos/src/screens/PassengerScreen/styles.js
@@ -1,6 +1,7 @@
 import {StyleSheet} from 'react-native';
 
 const styles = StyleSheet.create({
+  container: {height: '100%', width: '100%'},
   text: {
     marginLeft: 30,
     fontSize: 32,
